Modify the visibility of a lot of modules that should be private.
