+++ sparse/__init__.py
@@ -1,9 +1,8 @@
-from .coo import *
-from .compressed import GXCS
-from .dok import DOK
-from .sparse_array import SparseArray
-from .utils import random
-from .io import save_npz, load_npz
+from ._coo import *
+from ._dok import DOK
+from ._sparse_array import SparseArray
+from ._utils import random
+from ._io import save_npz, load_npz
 
 from ._version import get_versions
 __version__ = get_versions()['version']
