+++ sparse/_settings.py
