+++ sparse/_slicing.py
